move test into own folder
