+++ frontend/package.json
@@ -41,7 +41,6 @@
     "http-proxy-middleware": "^1.0.4",
     "ignore-loader": "^0.1.2",
     "jest": "^26.0.1",
-    "ts-loader": "^8.0.1",
     "typescript": "^3.9.3"
   },
   "lint-staged": {
